sign up page cleanup and basics
