+++ lib/app/app.dart
@@ -1,7 +1,7 @@
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_chatbot/app/models/theme_model.dart';
-import 'package:flutter_chatbot/app/services/firebase_signin.dart';
+import 'package:flutter_chatbot/app/services/firebase_auth_service.dart';
 import 'package:flutter_chatbot/ui/views/home/home_view.dart';
 import 'package:flutter_chatbot/ui/views/onboarding/onboard_pageview.dart';
 import 'package:flutter_chatbot/ui/views/private_route.dart';
